+++ lib/wexhook_web/controllers/page_html.ex
@@ -1,5 +1,0 @@
-defmodule WexhookWeb.PageHTML do
-  use WexhookWeb, :html
-
-  embed_templates "page_html/*"
-end