+++ test/wexhook_web/live/home_live_test.exs
@@ -5,19 +5,19 @@ defmodule WexhookWeb.HomeLiveTest do
   import Phoenix.LiveViewTest
 
   test "renders the page", %{conn: conn} do
-    {:ok, _view, html} = live(conn, "/")
+    {:ok, _view, html} = live(conn, ~p"/wexhook")
 
     assert html =~ "Webhook URL"
   end
 
   test "renders the page for a server fetching a server that doesn't exists", %{conn: conn} do
     id = "server_id"
-    conn = get(conn, "/share/" <> id)
+    conn = get(conn, ~p"/wexhook/share/#{id}")
     {:ok, _view, html} = live(conn)
 
     assert html =~ "Webhook URL"
 
-    conn = post(conn, "/hook/" <> id)
+    conn = post(conn, ~p"/wexhook/hook/#{id}")
     json_response(conn, 200)
 
     assert_receive {_, {:push_event, "request_received", %{}}}
@@ -27,12 +27,12 @@ defmodule WexhookWeb.HomeLiveTest do
     {:ok, server_pid} = Wexhook.new_server()
     id = Wexhook.get_server_id(server_pid)
 
-    conn = get(conn, "/share/" <> id)
+    conn = get(conn, ~p"/wexhook/share/#{id}")
     {:ok, _view, html} = live(conn)
 
     assert html =~ "Webhook URL"
 
-    conn = post(conn, "/hook/" <> id)
+    conn = post(conn, ~p"/wexhook/hook/#{id}")
     json_response(conn, 200)
 
     assert_receive {_, {:push_event, "request_received", %{}}}